refactor: update file location and function name
